Add Other.faa/non-Other.faa in ByVirus
We inserted the Other and non-Other data for ensuring project consistency. Please notice that the Other/non-Other sequences have been already integrated with negative datasets of different family/virus classes in the Fasta directory. Though it is trivial for the incoming model establishment, I still uploaded them for the integrity of the project.